objfx objParticleFn_80099d84 92.48->92.82 (int type, #11/#14)
The def's u8 type param made MWCC mask it (clrlwi r0,r30,24) before the
switch/cbuf[type*3] uses AND mis-prioritized the saved-reg coloring (current
kept light in r30 and the type web fell out). Widening to int type keeps it in
r30 across the fn (matching target's mr r30,r4) and drops the clrlwi (cmplwi
r30,8 direct). Synced the fx_800944A0_shared.h decl (its only caller of this fn
is objfx.c; the other 3 includers don't call it -> codegen-neutral). Remaining
~7pp: target frame -160 vs -144 + _savegpr helper-save form (more saved regs;
recipe #99/#67 class). Banked.

+++ include/main/dll/fx_800944A0_shared.h
@@ -353,7 +353,7 @@ void objfx_spawnLightPulse(void *obj, u8 type, int a3, u8 mode, void *light, f32
 void objfx_spawnFlaggedTrailBurst(void *obj, u8 mode, int p5, int p6, int p7, f32 fval);
 void projectileParticleFxFn_80099660(void *obj, int mode);
 void itemPickupDoParticleFx(void *obj, int mode, u8 count, f32 fval);
-void objParticleFn_80099d84(void *obj, f32 scale, u8 type, f32 fextra, void *light);
+void objParticleFn_80099d84(void *obj, f32 scale, int type, f32 fextra, void *light);
 void fn_8009A8C8(u8 *obj, f32 thresh);
 void DIMexplosionFn_8009a96c(u8 *src, f32 vx, f32 vy, f32 vz, f32 fval, u8 a, u8 flag4,
                              u8 flag8, u8 flag10, u8 doShake, u8 flag20, u8 f1cinit);
